Address review suggestions: move validate_device_requirements -> qjit_device.py
